feat(arch): PR 2/12 — move sqlite_base.py to infrastructure/stores
Move SqliteStore base class from core/stores to infrastructure/stores.
Update all 9 inheriting stores to import from the new location.
Update test file to use new import path.
Add ignore_imports for transitional core->infrastructure imports
(remaining stores pending migration in PRs 3-11).
Remove stale sqlite_base entries from forbidden contract ignore list.
Add turn_store.py to file length exempt list (pre-existing 301 lines).

Part of ADR-048 infrastructure layer rollout (#752).
