+++ src/lyra/core/stores/pairing.py
@@ -18,6 +18,7 @@ if TYPE_CHECKING:
     from lyra.core.stores.auth_store import AuthStore
 
 from lyra.core.trust import TrustLevel
+from lyra.infrastructure.stores.sqlite_base import SqliteStore
 
 from .pairing_config import (
     _CREATE_PAIRING_CODES,
@@ -27,7 +28,6 @@ from .pairing_config import (
     _sha256,
     _utc_now,
 )
-from .sqlite_base import SqliteStore
 
 # Re-export so existing `from lyra.core.pairing import PairingConfig/PairingError`
 # imports continue to work unchanged.
