+++ src/lyra/infrastructure/CLAUDE.md
@@ -15,7 +15,7 @@ All files will be moved from `src/lyra/core/stores/`:
 
 | File | Status | Destination |
 |------|--------|-------------|
-| `sqlite_base.py` | pending | `infrastructure/stores/sqlite_base.py` |
+| `sqlite_base.py` | migrated | `infrastructure/stores/sqlite_base.py` |
 | `agent_store.py` | pending | `infrastructure/stores/agent_store.py` |
 | `agent_store_migrations.py` | pending | `infrastructure/stores/agent_store_migrations.py` |
 | `auth_store.py` | pending | `infrastructure/stores/auth_store.py` |
