+++ src/lyra/infrastructure/stores/sqlite_base.py
